+++ packages/plugin-link-preview/web/src/typeMapper-loadable.ts
@@ -1,0 +1,7 @@
+import { LINK_PREVIEW_TYPE } from './types';
+import { PluginTypeMapper } from 'wix-rich-content-common';
+import loadable from '@loadable/component';
+
+export const typeMapper: PluginTypeMapper = () => ({
+  [LINK_PREVIEW_TYPE]: { component: loadable(() => import('./LinkPreviewViewer')) },
+});